refactor: change enemy package's name
